Add configs for path to write and read wacz archives
browsertrix_home_host is what gets passed to docker to mount on the
crawler container.

If auto-archiver is running on the host or we have a full
docker-in-docker situation, then this should be a local path and
defaults to a randomly generated folder.

If auto-archiver is running in docker, and wants to do docker out of
docker, this should be the name of a shared volume or a path on the
host.

browsertrix_home_container is where auto-archiver will look for
the wacz archives generated by the crawler.

If auto-archiver is running on the host or with full docker-in-docker then
this should default to the same as browsertrix_home_host

If auto-archiver is running in docker, and wants to do docker out of
docker, this should be path where the shared volume or host path is
mounted on the auto-archiver container.