Add environment variable to explicitly enable docker
When running from the host, auto-archiver will run the crawler in docker
by default. When running in docker, we default to calling crawl directly
to avoid an inadvertant docker-from-docker situation.

However we also want to support running docker from docker if explicitly
enabled (in a docker compose file, for example).

So the real question is not simply "am i in docker" but rather, "should
i call docker?"

This change cleans up the code a bit to reflect this and makes for a
more sensible configuration for docker-from-docker users who will set
WACZ_ENABLE_DOCKER=1 instead of the falsehood of  RUNNING_IN_DOCKER=0